Adding REST functions, changes in reddit_parser.py
Realized deletion a line from file with the identifier UNIQUE_ID;
Realized changing the content in line in file with identifier UNIQUE_ID;
Realized comparison of the hash sum of the file before changes
(deletions) and after to determine the response code;
Improved processing POST requests for data serialization in JSON format.

Changes in reddit_parser.py:

In the process of collecting information about users in the loop,
the function of sending a POST request to the server, which contains
a ready-made data dictionary, has been added.

Updated requirements.txt